requirements.txt
# Core data science stack
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Machine learning
lightgbm>=4.0.0
xgboost>=2.0.0

# Data storage and processing
pyarrow>=12.0.0
fastparquet>=2023.7.0

# API clients
praw>=7.7.0
requests>=2.31.0

# NLP and text processing
spacy>=3.6.0
vaderSentiment>=3.3.2
textstat>=0.7.3

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Statistical analysis
statsmodels>=0.14.0
pygam>=0.8.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0

# Development and testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0